refactor(skills): split default-layout/components.tsx per family (#40)
Phase 3 of the god-file cleanup. components.tsx (2,650 lines, 9 major
A2UI composites + markdown adapter + xterm theme helpers) is now a
39-line barrel that re-exports from focused sibling modules:

  markdown-adapter.tsx   MARKDOWN_COMPONENTS, HighlightedFence,
                         isHighlightedFenceChild
  layout.tsx             Layout, AeMarkInline, readUiScale, StatusBar,
                         EmptyState
  sidebar/               filter.ts (providerOf, filterItems),
                         item-row.tsx, searchable-section.tsx,
                         index.tsx (Sidebar)
  chat.tsx               ChatHistory, ChatInput + slash palette,
                         ToolCard, MainCanvas, formatToolDuration,
                         roleBadge, ScrollToBottomPill
  terminal.tsx           Terminal + readTerminalTheme +
                         observeTerminalTheme (shared with shell/)
  shell/                 canvas.tsx (ShellCanvas + ShellStatusBar),
                         panel.tsx (TerminalPanel + SubTabPill),
                         tab-strip.tsx (TabStrip)

No behavior change. index.ts and components.test.ts imports unchanged.
All 305 vitest + 73 cargo tests pass; lint, tsc, clippy clean.
