+++ src/skills/default-layout/shell/index.ts
@@ -1,0 +1,3 @@
+export { ShellCanvas } from "./canvas";
+export { TerminalPanel } from "./panel";
+export { TabStrip } from "./tab-strip";
